requirements.txt
groq
python-dotenv
pillow
numpy
imageio
imageio-ffmpeg
moviepy
scipy
gtts
edge-tts
requests
pytrends
flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
